LayoutTests/js/stringimpl-to-jsstring-on-large-strings-1-expected.txt

PASS createStrings() threw exception RangeError: Out of memory.
PASS successfullyParsed is true

TEST COMPLETE
